change config init method name
