+++ blockchain/browsersync/bsc/bscSetup.go
@@ -1,8 +1,9 @@
-package bscclient
+package bsc
 
 import (
+	"context"
 	"github.com/ethereum/go-ethereum/ethclient"
-	"payment-bridge/config"
+	"math/big"
 	"payment-bridge/logs"
 	"time"
 )
@@ -16,7 +17,7 @@ var WebConn = new(ConnSetup)
 
 func ClientInit() {
 	for {
-		rpcUrl := config.GetConfig().BscMainnetNode.RpcUrl
+		rpcUrl := GetConfig().BscMainnetNode.RpcUrl
 		client, err := ethclient.Dial(rpcUrl)
 		if err != nil {
 			logs.GetLogger().Error("Try to reconnect block chain node" + rpcUrl + " ...")
@@ -26,4 +27,17 @@ func ClientInit() {
 			break
 		}
 	}
+}
+
+func (conn *ConnSetup) GetBlockNumber() (*big.Int, error) {
+	//return conn.ConnWeb.Eth.GetBlockNumber()
+	block, err := conn.ConnWeb.HeaderByNumber(context.Background(), nil)
+	if block != nil {
+		return block.Number, err
+	}
+	return nil, err
+}
+
+func init() {
+	ClientInit()
 }
